Add formatter module and keyword test
